+++ cairo/scripts/__init__.py
@@ -1,0 +1,4 @@
+# Apply EELS types patches before importing any other modules
+from keth_types.patches import apply_patches
+
+apply_patches()
